Some types to Type.ts files
