+++ src/utils/function.ts
@@ -1,6 +1,6 @@
-export const isTruthyObj = (obj: object) => {
-  return !!Object.keys(obj).length;
-};
+// export const isTruthyObj = (obj: object) => {
+//   return !!Object.keys(obj).length;
+// };
 
 export function api<T>(url: string): Promise<T> {
   return fetch(url).then(response => {
